% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexity}
\alias{complexity}
\title{Complexity Index}
\usage{
complexity(RCAmat, measure = c("eci", "pci"))
}
\arguments{
\item{RCAmat}{An economy-by-product matrix with the RCA values.}

\item{measure}{A character string specifying the measure to compute:
\enumerate{
\item "eci": economic complexity index
\item "pci": product complexity index
}}
}
\value{
A numeric vector with the measured values.
}
\description{
Compute the economic complexity index (ECI) and the product complexity index (PCI)
\insertCite{hidalgo2009building}{ProductSpace} based on the RCA values.
}
\references{
\insertAllCited{}
}
