% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density}
\alias{density}
\title{Density}
\usage{
density(RCAmat, PROXmat)
}
\arguments{
\item{RCAmat}{An economy-by-product matrix with the RCA values.}

\item{PROXmat}{A product-by-product matrix with the proximity values.}
}
\value{
A economy-by-product matrix with the density values.
}
\description{
Compute the density \insertCite{hausmann2006structural}{ProductSpace} based on the RCA
and proximity values.
}
\references{
\insertAllCited{}
}
