% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance}
\usage{
distance(RCAmat, PROXmat)
}
\arguments{
\item{RCAmat}{An economy-by-product matrix with the RCA values.}

\item{PROXmat}{A product-by-product matrix with the proximity values.}
}
\value{
A economy-by-product matrix with the distance values.
}
\description{
Compute the distance \insertCite{hausmann2006structural}{ProductSpace} based on the RCA
and proximity values.
}
\references{
\insertAllCited{}
}
