% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Diversity}
\usage{
diversity(RCAmat)
}
\arguments{
\item{RCAmat}{An economy-by-product matrix with the RCA values.}
}
\value{
A numeric vectors.
}
\description{
Compute the diversity/diversification \insertCite{hidalgo2009dynamics}{ProductSpace}
based on RCA values.
}
\references{
\insertAllCited{}
}
