% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{ubiquity}
\alias{ubiquity}
\title{Ubiquity}
\usage{
ubiquity(RCAmat)
}
\arguments{
\item{RCAmat}{An economy-by-product matrix with the RCA values.}
}
\value{
A numeric vectors.
}
\description{
Compute the ubiquity \insertCite{hidalgo2009dynamics}{ProductSpace} based on RCA values.
}
\references{
\insertAllCited{}
}
